requirements.txt
streamlit
pandas
openpyxl
scikit-learn
openai
google-generativeai
python-dotenv
pdfkit
plotly
